dataplane/userspace: split zones.go by concern (#4421)
Pure code-motion refactor of pkg/dataplane/userspace/zones.go (1222
LOC, 23 top-level decls fusing five unrelated concerns) into the
residual file plus five sibling files, one per concern. No behavior
change: every top-level declaration moves verbatim, the imports are
re-derived per file, and gofmt is applied. The sorted production
decl-name set is unchanged and the gofmt'd body multiset is identical
(1181 non-blank content lines, imports excluded).

Concern-to-file mapping:

  zones.go (residual) — helpers shared across >=2 concerns:
    hostInboundLifelineSet / hostInboundLifelineInterface (kernel
    lifeline wrappers), hostIPFromCIDR, buildInterfaceZoneMap.

  zones_host_inbound.go — kernel-nft host-inbound enforcement views:
    ZoneHostInboundView, BuildZoneHostInboundViews,
    UnzonedHostInboundZoneLabel, BuildUnzonedHostInboundAddrs.

  zones_observability.go — fail-open / ambiguity reporters that read
    back from the enforcement builder: AddresslessEnforcingZone(s),
    AddresslessDHCPPending, AddresslessEnforcingInterface(s),
    AmbiguousHostInboundAddress(es).

  zones_override.go — per-interface host-inbound override resolution:
    unionHostInboundTokens, mergeHostInboundTraffic,
    buildInterfaceHostInboundMap.

  zones_snapshot.go — zone-ID / wire-snapshot construction:
    buildZoneSnapshots, lowerTokens.

  zones_quarantine.go — StableZoneID collision quarantine:
    ZoneIDCollision (+String), quarantineCollidingZones.

buildZoneSnapshots (used by builder.go) and buildInterfaceZoneMap
(used by interfaces.go) are the only cross-file consumers; all other
symbols are package-internal, so the motion is transparent to callers.

Validation: go build ./... green; go test
./pkg/dataplane/userspace/... green (13.3s); decl-name-set empty diff
vs origin/master; gofmt body multiset byte-identical.
