+++ pkg/dataplane/userspace/zones_host_inbound.go
@@ -1,0 +1,394 @@
+package userspace
+
+import (
+	"fmt"
+	"sort"
+	"strings"
+
+	"github.com/psaab/xpf/pkg/config"
+)
+
+// ZoneHostInboundView is the per-zone host-inbound-traffic enforcement view for
+// the KERNEL-nftables primary path (#3070). Ordinary host-bound traffic to a
+// firewall interface IP / VRRP VIP (SSH, ping, OSPF/BGP to the box) is shunted
+// to the Linux kernel by the XDP shim before it ever reaches userspace-dp, so
+// the authoritative host-inbound enforcement for those packets must live in the
+// kernel `chain input` (mirroring the lo0-filter precedent). The userspace-dp
+// LocalDelivery check (forwarding/host_inbound.rs) remains the secondary path
+// for the narrow subset that DOES reach the XSK (DNAT-to-self, static-NAT to a
+// firewall service, embedded-ICMP, DNS edge cases).
+//
+// One view is produced per host-inbound-CONFIGURED zone, carrying the zone's
+// allowed Junos tokens plus its resolved firewall-local host addresses (bare
+// IPs, prefix stripped), split by family. The daemon maps the tokens to nft
+// matches and emits accept-the-listed / deny-the-rest rules scoped to those
+// addresses.
+type ZoneHostInboundView struct {
+	Zone string
+	// Interfaces lists the interface refs whose EFFECTIVE host-inbound token
+	// set (zone-level ∪ interface-level override, #3362) equals this view's
+	// SystemServices/Protocols. A zone with no per-interface override yields a
+	// single view per zone covering all its interfaces (pre-#3362 shape); a zone
+	// with an override yields one view per distinct effective token set, each
+	// scoped to that set's interface addresses. Sorted; informational/test only
+	// (the nft emission keys on the address set, which is per-interface).
+	Interfaces     []string
+	SystemServices []string
+	Protocols      []string
+	V4Addrs        []string // bare host IPv4 addresses (no prefix)
+	V6Addrs        []string // bare host IPv6 addresses (no prefix)
+}
+
+// BuildZoneHostInboundViews returns one ZoneHostInboundView per configured
+// security zone (#3070; #3405 default-deny parity — every zone enforces, see
+// below), resolving each zone's firewall-local
+// host addresses via the canonical interface-snapshot builder (the same
+// resolution that populates the dataplane) PLUS each zone's RETH VRRP virtual
+// addresses (#3172, resolved from config so they scope the deny on the backup
+// node too, where the VIP is not yet live on the kernel interface), with
+// management/cluster-control lifeline interfaces (fxp0 / em0 / fab*) excluded
+// from the address set.
+//
+// Address completeness (#3224 — non-reproducing): the snapshot builder resolves
+// each interface's addresses through buildLinkSnapshot -> AddrList(FAMILY_ALL),
+// which enumerates EVERY kernel address with no scope/flag/dynamic filtering.
+// So DHCP / DHCPv6-learned addresses are captured exactly like static ones —
+// a DHCP-only interface with a live lease yields a NON-empty address set and IS
+// scoped by the deny. (xpfd disables IPv6 RA on every managed interface in
+// pkg/networkd, so DHCPv6 is the only IPv6 dynamic-address path and the same
+// snapshot captures it; SLAAC is not a separate case.) The deny is also
+// re-rendered on every DHCP/DHCPv6 lease change on a dataplane interface
+// (onDHCPAddressChange -> dhcpLeaseChangeRequiresRecompile -> applyConfig ->
+// applyHostInboundFilter), so a renewed/flapped lease re-scopes within one
+// reconcile pass. #3224 was filed on the premise that DHCP addresses fell out
+// of scope (FAIL OPEN); that does not reproduce because the live snapshot has
+// always carried them — see TestBuildZoneHostInboundViewsScopesKernelLearnedAddr,
+// which exercises this real path with a config-absent kernel address.
+//
+// #3405: a zone that declared NO host-inbound-traffic stanza is NOT omitted —
+// it is treated as an empty stanza and gets a catch-all DROP scoped to its
+// firewall-local addresses (Junos default-deny: deny every host-bound
+// service/protocol not explicitly permitted). The only no-address case left is a
+// configured zone whose interfaces have neither a static config address nor any
+// live kernel address yet (e.g. a DHCP WAN before its first lease, or a backup
+// node before VIP install): it yields an empty address set, the daemon emits no
+// deny for it, and it self-heals once an address appears because the
+// lease-change / commit paths re-render. That transient fail-open admit window
+// is surfaced to operators by AddresslessEnforcingZones (#3698) — the daemon
+// logs a state-transition warning and exports xpf_host_inbound_addressless_zones
+// while the window is open, so it is no longer silent.
+func BuildZoneHostInboundViews(cfg *config.Config) []ZoneHostInboundView {
+	if cfg == nil || len(cfg.Security.Zones) == 0 {
+		return nil
+	}
+	ifaceSnaps := buildInterfaceSnapshots(cfg)
+	// Lifeline interfaces (fxp0 + the configured chassis-cluster
+	// control-interface / fabric interfaces, plus the em0/fab* defaults) are
+	// excluded from host-inbound deny scoping so management / cluster-control
+	// traffic is never denied (#3277).
+	lifelines := hostInboundLifelineSet(cfg)
+	// #3362: per-interface host-inbound override lookup (ref → override, with
+	// physical→unit expansion). The EFFECTIVE token set for an interface is the
+	// UNION of its zone-level set and this override; interfaces in the same zone
+	// with the SAME effective set share one view (one nft address set), so a zone
+	// with NO override produces exactly one view per zone (pre-#3362 shape).
+	overrideByIface := buildInterfaceHostInboundMap(cfg)
+
+	// Each emitted view is a group keyed by (zone, effective-token signature).
+	// Addresses accumulate per group; a group is created lazily on its first
+	// address so a configured-but-address-less interface stays omitted (admit
+	// nothing emitted, self-heals when an address appears) exactly as before.
+	type group struct {
+		zone   string
+		svc    []string
+		proto  []string
+		v4, v6 []string
+		seen4  map[string]bool
+		seen6  map[string]bool
+		ifaces map[string]bool
+	}
+	groups := make(map[string]*group)
+	getGroup := func(zone string, svc, proto []string, iface string) *group {
+		// #3721: group by a CANONICAL (sorted, deduped) token signature so two
+		// interfaces whose EFFECTIVE admission sets are semantically identical but
+		// authored in a different order ([ssh ping] vs [ping ssh]) fall into ONE
+		// group — one nft rule block + one deny counter — instead of the
+		// order-sensitive strings.Join keying two groups that inflate the nft
+		// payload / `nft -f` replace time on a large trunk. The group keeps the
+		// FIRST-seen authored svc/proto order for display fidelity; enforcement
+		// keys on the address set plus the accept-token set, both
+		// order-independent, so this is behavior-preserving (identical admission,
+		// fewer duplicate blocks). Shares config.CanonicalHostInboundTokenSig with
+		// the commit gate and the ambiguity reporter so all three agree on what
+		// counts as "the same set".
+		sig := zone + "\x00" + config.CanonicalHostInboundTokenSig(svc, proto)
+		g := groups[sig]
+		if g == nil {
+			g = &group{
+				zone: zone, svc: svc, proto: proto,
+				seen4: map[string]bool{}, seen6: map[string]bool{},
+				ifaces: map[string]bool{},
+			}
+			groups[sig] = g
+		}
+		if iface != "" {
+			g.ifaces[iface] = true
+		}
+		return g
+	}
+	addAddr := func(g *group, host string) {
+		if strings.Contains(host, ":") {
+			if !g.seen6[host] {
+				g.seen6[host] = true
+				g.v6 = append(g.v6, host)
+			}
+		} else if !g.seen4[host] {
+			g.seen4[host] = true
+			g.v4 = append(g.v4, host)
+		}
+	}
+	// configured reports whether a zone enforces host-inbound at all. #3405:
+	// EVERY configured security zone enforces host-inbound (Junos/vSRX
+	// default-deny parity) — a zone with interfaces but NO `host-inbound-traffic`
+	// stanza is treated identically to an empty stanza (`host-inbound-traffic
+	// { }`): it scopes a catch-all kernel DROP to its firewall-local addresses,
+	// denying every host-bound service/protocol not explicitly permitted. Before
+	// #3405 a no-stanza zone was skipped entirely (admit-all), a permit-all
+	// management-plane exposure on any zone the operator never locked down. The
+	// management / cluster-control lifeline interfaces (fxp0/em0/fab*) are
+	// excluded from the address sets below, and the established / ESP-AH / ND /
+	// PMTUD accepts precede every drop, so the default-deny can never strand
+	// management or break HA. A zone-level stanza or any per-interface override
+	// (#3362) still further scopes what the zone admits.
+	configured := func(zone *config.ZoneConfig) bool {
+		return zone != nil
+	}
+
+	// Seed a zone-default group (zone-level effective tokens, no override) for
+	// every configured zone so each configured zone yields at least one view —
+	// even when its only interface is a lifeline (no address contributed) — the
+	// pre-#3362 "one view per configured zone" contract. Non-overridden
+	// interfaces accumulate their addresses into this same group (identical
+	// signature); fully-overridden / address-less zones keep an empty view (the
+	// daemon emits no deny for it).
+	zoneNamesSorted := make([]string, 0, len(cfg.Security.Zones))
+	for name := range cfg.Security.Zones {
+		zoneNamesSorted = append(zoneNamesSorted, name)
+	}
+	sort.Strings(zoneNamesSorted)
+	for _, name := range zoneNamesSorted {
+		zone := cfg.Security.Zones[name]
+		if !configured(zone) {
+			continue
+		}
+		svc, proto := unionHostInboundTokens(zone.HostInboundTraffic, nil)
+		getGroup(name, svc, proto, "")
+	}
+
+	// Per-interface static/learned addresses from the resolved snapshots.
+	for _, snap := range ifaceSnaps {
+		if snap.Zone == "" || hostInboundLifelineInterface(snap.Name, lifelines) {
+			continue
+		}
+		zone := cfg.Security.Zones[snap.Zone]
+		if !configured(zone) {
+			continue
+		}
+		svc, proto := unionHostInboundTokens(zone.HostInboundTraffic, overrideByIface[snap.Name])
+		var g *group
+		for _, a := range snap.Addresses {
+			host := hostIPFromCIDR(a.Address)
+			if host == "" {
+				continue
+			}
+			if g == nil {
+				g = getGroup(snap.Zone, svc, proto, snap.Name)
+			}
+			addAddr(g, host)
+		}
+	}
+
+	// VRRP RETH VIPs (#3172): the host-inbound destination address set must
+	// also include each zone's RETH virtual IPs, not only the static interface
+	// addresses resolved above. A VIP is present on the kernel interface ONLY of
+	// the node that currently owns the redundancy group (master); on the backup
+	// node the VIP is absent from buildLinkSnapshot's live address list, so
+	// without this the kernel host-inbound deny would not be scoped to the VIP
+	// and `chain input` would fall through to `policy accept` (FAIL-OPEN) for
+	// VIP-destined host-bound traffic. The VIPs are identical on both nodes, so
+	// resolving them from config (unit.VRRPGroups[*].VirtualAddresses) scopes the
+	// deny consistently regardless of mastership. The seen maps dedup against the
+	// live snapshot, so on the master node (where the VIP is already live) the
+	// result is byte-identical. Lifeline interfaces (fxp0/em0/fab*) are excluded,
+	// mirroring the static-address path; standalone (no-VRRP) zones are untouched.
+	// The VIP is added to its interface's EFFECTIVE-token group (#3362), so a VIP
+	// on an overridden interface is scoped by that interface's override.
+	zoneByIface := buildInterfaceZoneMap(cfg)
+	ifNames := make([]string, 0, len(cfg.Interfaces.Interfaces))
+	for n := range cfg.Interfaces.Interfaces {
+		ifNames = append(ifNames, n)
+	}
+	sort.Strings(ifNames)
+	for _, ifName := range ifNames {
+		iface := cfg.Interfaces.Interfaces[ifName]
+		if iface == nil {
+			continue
+		}
+		unitNums := make([]int, 0, len(iface.Units))
+		for u := range iface.Units {
+			unitNums = append(unitNums, u)
+		}
+		sort.Ints(unitNums)
+		for _, un := range unitNums {
+			unit := iface.Units[un]
+			if unit == nil || len(unit.VRRPGroups) == 0 {
+				continue
+			}
+			unitName := fmt.Sprintf("%s.%d", ifName, un)
+			if hostInboundLifelineInterface(unitName, lifelines) {
+				continue
+			}
+			zoneName := zoneByIface[unitName]
+			if zoneName == "" {
+				continue
+			}
+			zone := cfg.Security.Zones[zoneName]
+			if !configured(zone) {
+				continue
+			}
+			svc, proto := unionHostInboundTokens(zone.HostInboundTraffic, overrideByIface[unitName])
+			vgKeys := make([]string, 0, len(unit.VRRPGroups))
+			for k := range unit.VRRPGroups {
+				vgKeys = append(vgKeys, k)
+			}
+			sort.Strings(vgKeys)
+			for _, k := range vgKeys {
+				vg := unit.VRRPGroups[k]
+				if vg == nil {
+					continue
+				}
+				for _, vip := range vg.VirtualAddresses {
+					if host := hostIPFromCIDR(vip); host != "" {
+						addAddr(getGroup(zoneName, svc, proto, unitName), host)
+					}
+				}
+			}
+		}
+	}
+
+	// Emit groups deterministically: the signature begins with the zone name,
+	// so sorting by signature orders views by zone then by token set. Addresses
+	// within a view are sorted for a reproducible nft payload.
+	sigs := make([]string, 0, len(groups))
+	for sig := range groups {
+		sigs = append(sigs, sig)
+	}
+	sort.Strings(sigs)
+	out := make([]ZoneHostInboundView, 0, len(sigs))
+	for _, sig := range sigs {
+		g := groups[sig]
+		ifaces := make([]string, 0, len(g.ifaces))
+		for name := range g.ifaces {
+			ifaces = append(ifaces, name)
+		}
+		sort.Strings(ifaces)
+		// Addresses are kept in accumulation order (static interface addresses
+		// first, then VRRP VIPs) — NOT sorted — to preserve the pre-#3362 nft
+		// payload ordering. Order is deterministic: snapshots come from
+		// sorted-name iteration and VIPs from sorted interface/unit/group walks.
+		out = append(out, ZoneHostInboundView{
+			Zone:           g.zone,
+			Interfaces:     ifaces,
+			SystemServices: g.svc,
+			Protocols:      g.proto,
+			V4Addrs:        g.v4,
+			V6Addrs:        g.v6,
+		})
+	}
+	return out
+}
+
+// UnzonedHostInboundZoneLabel is the sentinel zone label under which the kernel
+// host-inbound catch-all deny for firewall-local addresses on interfaces
+// assigned to NO security zone is counted (#4420 HI-2). It reuses the reserved
+// Junos self-traffic context token "junos-host": that token can NEVER name an
+// operator-defined security zone (validateReservedZoneNamesStrict rejects it),
+// so the nft named-counter object it yields (nftables.HostInboundDenyCounterName)
+// can never collide with a real per-zone deny counter, and the #3361 scraper
+// (ParseHostInboundDenyCounterName) recovers a stable, self-explanatory
+// zone="junos-host" label for these "traffic to the host with no source zone"
+// host-inbound drops.
+const UnzonedHostInboundZoneLabel = "junos-host"
+
+// BuildUnzonedHostInboundAddrs returns the firewall-local host addresses (bare
+// IPs, prefix stripped, split by family) of interfaces that carry an address but
+// are assigned to NO security zone (#4420 HI-2).
+//
+// xpfd applies an interface's configured / leased address regardless of zone
+// membership (pkg/dataplane/compiler_iface.go builds the networkd managed set
+// from cfg.Interfaces, not from zones), yet BuildZoneHostInboundViews scopes the
+// kernel host-inbound default-deny ONLY to ZONED addresses. The kernel
+// `xpf_hostinbound` chain runs with `policy accept`, so host-bound traffic to an
+// addressed-but-unzoned interface falls through to accept — the host stack is
+// exposed on it with no host-inbound admission. That is a fail-open, and a
+// deviation from Junos, where an interface not in a security zone passes no
+// flow / host-inbound traffic at all. This builder collects those addresses so
+// the daemon can emit a catch-all DROP scoping them, restoring the Junos
+// fail-closed posture and mirroring the #3405 per-zone default-deny.
+//
+// Scope / safety:
+//   - Only meaningful when the operator uses the zone model at all (>= 1 zone):
+//     a zone-less bootstrap / degenerate config is left untouched (nil), so this
+//     never turns a no-zones box into deny-all host-inbound.
+//   - Management / cluster-control LIFELINE interfaces (fxp0 / em0 / fab*, plus
+//     the configured control / fabric links) are excluded exactly as the zone
+//     path excludes them, so the deny can never strand management or break HA.
+//   - Addresses already scoped by a zone view are subtracted, so a (mis)config
+//     placing one firewall-local address on both a zoned and an unzoned
+//     interface never yields a duplicate / conflicting rule for the same daddr.
+//   - Unzoned interfaces are NOT AF_XDP-bound (only zoned dataplane interfaces
+//     get the shim), so their host-bound traffic is delivered entirely through
+//     the kernel; the kernel nft deny is the sole and sufficient enforcement
+//     point and no userspace-dp (AF_XDP) change is required.
+func BuildUnzonedHostInboundAddrs(cfg *config.Config) (v4, v6 []string) {
+	if cfg == nil || len(cfg.Security.Zones) == 0 || len(cfg.Interfaces.Interfaces) == 0 {
+		return nil, nil
+	}
+	lifelines := hostInboundLifelineSet(cfg)
+	// Addresses already covered by a zone deny — exclude so the unzoned catch-all
+	// never duplicates or conflicts with a zone rule for the same daddr.
+	zoned := map[string]bool{}
+	for _, view := range BuildZoneHostInboundViews(cfg) {
+		for _, a := range view.V4Addrs {
+			zoned[a] = true
+		}
+		for _, a := range view.V6Addrs {
+			zoned[a] = true
+		}
+	}
+	seen4 := map[string]bool{}
+	seen6 := map[string]bool{}
+	for _, snap := range buildInterfaceSnapshots(cfg) {
+		if snap.Zone != "" || hostInboundLifelineInterface(snap.Name, lifelines) {
+			continue
+		}
+		for _, a := range snap.Addresses {
+			host := hostIPFromCIDR(a.Address)
+			if host == "" || zoned[host] {
+				continue
+			}
+			if strings.Contains(host, ":") {
+				if !seen6[host] {
+					seen6[host] = true
+					v6 = append(v6, host)
+				}
+			} else if !seen4[host] {
+				seen4[host] = true
+				v4 = append(v4, host)
+			}
+		}
+	}
+	sort.Strings(v4)
+	sort.Strings(v6)
+	return v4, v6
+}
